Set up Travis CI.
This also means making the tests a bit more robust: we now arrange for
git-rbr and git-rbr-core to be on the PATH, rather than depend on that
already being the case from the environment.

We use the `trusty` environment in Travis.  It's still in beta, which
is a bit awkward, but then again Precise is going unsupported upstream
next week.  Travis apparently doesn't plan to switch the default until
"sometime in July": https://blog.travis-ci.com/2017-04-17-precise-EOL
but hey, let's live in the future.